[New] Pattern editor: Channel rename dialog in channel header context menu [Fix] Mod Specs: The XM and MOD format does not support "---" or "+++" items in the order list. "+++" and "---" are left out when saving as XM. The XM format supports up to 255 patterns in theory, but pattern 254/255 are still loaded as "+++"/"---" for now. [Fix] General tab: Channel name input field was not limited properly. [Imp] Orderlist: Middle click can now also be used for queuing patterns. [Imp] Orderlist: "+++" pattern item cannot be entered when writing an MOD / XM file.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@413 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_med.cpp
@@ -636,7 +636,7 @@ bool CSoundFile::ReadMed(const BYTE *lpStream, DWORD dwMemLength)
 	} else
 	{
 		UINT nOrders, nSections;
-		UINT nTrks = BigEndianW(pmsh2->numtracks);
+		WORD nTrks = BigEndianW(pmsh2->numtracks);
 		if ((nTrks >= 4) && (nTrks <= 32)) m_nChannels = nTrks;
 		DWORD playseqtable = BigEndian(pmsh2->playseqtable);
 		UINT numplayseqs = BigEndianW(pmsh2->numpseqs);
